+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 Änderungen
 ----------
 v1.0.16 (??.??.2016)
+ * File permissions fixed for lock files
  * [BREAKING CHANGE] class properties tx_rnbase_configurations have now a visibility. Thus access e.g. to _dataStore will now fail.
 
 v1.0.15 (20.06.2016)